Move labelSelector schema and its flatteners/expanders functions out of persistant volume claims files to their own files for reuse
